Factor out DisplayMode
